Move code to right place
