+++ Sources/RequestDL/Internals/Sources/Client/Client/Internals.Client.swift
@@ -60,7 +60,7 @@ extension Internals {
                 delegate: delegate
             )
 
-            return UnsafeTask(task) { _ in
+            return UnsafeTask(task) {
                 _Concurrency.Task {
                     await self.lock.withLockVoid {
                         operation.complete()
@@ -79,117 +79,6 @@ extension Internals {
                 _isClosed = true
                 return true
             }
-        }
-    }
-}
-
-extension Internals {
-
-    enum TaskStatus {
-        case finished
-        case cancelled
-    }
-
-    final class TaskSeed: Sendable, Hashable {
-
-        static var withoutCancellation: TaskSeed {
-            TaskSeed {}
-        }
-
-        private let cancel: @Sendable () -> Void
-
-        init(_ cancel: @escaping @Sendable () -> Void) {
-            self.cancel = cancel
-        }
-
-        static func == (lhs: Internals.TaskSeed, rhs: Internals.TaskSeed) -> Bool {
-            lhs === rhs
-        }
-
-        func hash(into hasher: inout Hasher) {
-            hasher.combine(ObjectIdentifier(self))
-        }
-
-        @Sendable
-        func callAsFunction() {
-            cancel()
-        }
-
-        deinit {
-            cancel()
-        }
-    }
-
-    struct UnsafeTask<Element>: Sendable, Hashable {
-
-        fileprivate final class RunningState: @unchecked Sendable {
-
-            var isRunning: Bool {
-                get { lock.withLock { _isRunning } }
-                set { lock.withLock { _isRunning = newValue } }
-            }
-
-            private let lock = Lock()
-
-            private var _isRunning = true
-        }
-
-        // MARK: - Private properties
-
-        private let task: HTTPClient.Task<Element>
-        private let seed: TaskSeed
-
-        // MARK: - Inits
-
-        init(
-            _ task: HTTPClient.Task<Element>,
-            completion: @escaping (TaskStatus) -> Void
-        ) {
-            let runningState = RunningState()
-
-            seed = TaskSeed {
-                guard runningState.isRunning else {
-                    return
-                }
-
-                runningState.isRunning = false
-                task.cancel()
-                completion(.cancelled)
-            }
-
-            task.futureResult.whenComplete { _ in
-                guard runningState.isRunning else {
-                    return
-                }
-
-                runningState.isRunning = false
-                completion(.finished)
-            }
-
-            self.task = task
-        }
-
-        // MARK: - Internal static methods
-
-        static func == (_ lhs: Self, _ rhs: Self) -> Bool {
-            lhs.seed === rhs.seed
-        }
-
-        // MARK: - Internal methods
-
-        func response() async throws -> Element {
-            try await withTaskCancellationHandler(
-                operation: task.futureResult.get,
-                onCancel: seed.callAsFunction
-            )
-        }
-
-        func callAsFunction() -> TaskSeed {
-            seed
-        }
-
-        func hash(into hasher: inout Hasher) {
-            hasher.combine(ObjectIdentifier(seed))
         }
     }
 }
